% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIID-package.R
\docType{package}
\name{BIID-package}
\alias{BIID-package}
\title{BIID: a Package for performing efficient Bayesian inference for infectious disease models.}
\description{
Efficient Bayesian Inference Methods for Infectious Disease Models.
}
\details{
The main function of the package is:

 \describe{ \item{MCMCiFFBS_}{Function that performs MCMC with iFFBS}}.
}
\author{
Evandro Konzen, Trevelyan J. McKinley.
}
