% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MCMCiFFBS_}
\alias{MCMCiFFBS_}
\title{Performs MCMC-iFFBS for CMR-Gompertz-SEID-rho model with group-specific alpha}
\usage{
MCMCiFFBS_(
  N,
  initParamValues,
  Xinit,
  TestMat,
  CaptHist,
  birthTimes,
  startSamplingPeriod,
  endSamplingPeriod,
  nuTimes,
  CaptEffort,
  capturesAfterMonit,
  numSeasons,
  seasonStart,
  maxt,
  hp_lambda,
  hp_beta,
  hp_q,
  hp_tau,
  hp_a2,
  hp_b2,
  hp_c1,
  hp_nu,
  hp_xi,
  hp_theta,
  hp_rho,
  hp_phi,
  hp_eta,
  k,
  K,
  sd_xi_min,
  method,
  epsilon,
  epsilonalphas,
  epsilonbq,
  epsilontau,
  epsilonc1,
  epsilonsens,
  L,
  path,
  blockSize
)
}
\arguments{
\item{N}{Number of MCMC iterations}

\item{initParamValues}{Initial parameter values for the two infection rates;
rate at latent period; three Gompertz parameters;
initial probability of being susceptible, exposed, and infectious;
test sensitivities; rho scaling parameters; test specificities; and 
seasonal capture probabilities. If set to Inf, then the 
initial parameter values will be generated from their priors.}

\item{Xinit}{Matrix of dimension (\code{m} x \code{maxt}) containing the initial states}

\item{TestMat}{Matrix with all capture events. 
The columns are (time, id, group, test1, test2, test3, ...).}

\item{CaptHist}{Matrix of dimension (\code{m} x \code{maxt}) containing the capture history}

\item{birthTimes}{Integer vector of length \code{m} representing the birth times}

\item{startSamplingPeriod}{Integer vector of length \code{m} with the start sampling times}

\item{endSamplingPeriod}{Integer vector of length \code{m} with the end sampling times}

\item{nuTimes}{Integer vector saying at which times nu parameters are applied}

\item{CaptEffort}{(G x maxt) matrix with 1 indicates monitoring and 0 otherwise}

\item{capturesAfterMonit}{matrix containing the last capture times for individuals captured after the monitoring period}

\item{numSeasons}{Number of seasons}

\item{seasonStart}{Starting season}

\item{maxt}{Number of time points}

\item{hp_lambda}{Vector of hyperparameter values for the Gamma prior on mean of alpha (1/lambda)}

\item{hp_beta}{Vector of hyperparameter values for the Gamma prior on frequency-dependent transmission rate}

\item{hp_q}{Vector of hyperparameter values for the Gamma prior on q}

\item{hp_tau}{Vector of hyperparameter values for the Gamma prior on average latent period}

\item{hp_a2}{Vector of hyperparameter values for the Gamma prior on Gompertz parameter a2}

\item{hp_b2}{Vector of hyperparameter values for the Gamma prior on Gompertz parameter b2}

\item{hp_c1}{Vector of hyperparameter values for the Gamma prior on Gompertz parameter c1}

\item{hp_nu}{Vector of hyperparameter values for Dirichlet prior on the 
initial probability of infection of being susceptible, exposed, infectious}

\item{hp_xi}{Vector of hyperparameter values (mean, std deviation) for the prior on the Brock changepoint}

\item{hp_theta}{Vector of hyperparameter values for the Beta prior on test 
sensitivities}

\item{hp_rho}{Vector of hyperparameter values for the Beta prior on scaling 
factor of test sensitivities in the latent period}

\item{hp_phi}{Vector of hyperparameter values for the Beta prior on test 
specificities}

\item{hp_eta}{Vector of hyperparameter values for the Beta prior on capture 
probabilities}

\item{k}{Positive integer (shape) parameter of Gamma distribution for the latent period}

\item{K}{Rescaling parameter for the population size (in order to have beta independent of q)}

\item{sd_xi_min}{Minimum value for the proposal standard deviation}

\item{method}{Integer indicating which method is used for updating 
infection rates and Gompertz parameters (method=1 for "HMC" and method=2 for "RWMH")}

\item{epsilon}{Leapfrog stepsize in HMC}

\item{epsilonalphas}{Leapfrog stepsize in HMC for alphas}

\item{epsilonbq}{Leapfrog stepsize in HMC for beta and q}

\item{epsilontau}{Leapfrog stepsize in HMC for tau}

\item{epsilonc1}{Leapfrog stepsize in HMC for c1}

\item{epsilonsens}{Leapfrog stepsize in HMC for thetas and rhos}

\item{L}{Number of leapfrog steps in HMC}

\item{path}{Directory's name where results will be saved on}

\item{blockSize}{number of iterations saved in each folder}
}
\value{
A matrix containing MCMC draws (\code{N} rows) from
the posterior distribution of all parameters.//'
}
\description{
Performs MCMC-iFFBS for CMR-Gompertz-SEID-rho model with group-specific alpha
}
\details{
\itemize{ \item Columns having test results in \code{TestMat} 
should have values
\itemize{ 
\item \code{0}: negative test 
\item \code{1}: positive test 
\item \code{NA}: test not observed
}
\item \code{Capture history} should have values
\itemize{
\item \code{0}: not captured 
\item \code{1}: captured
}
\item \code{Xinit} should have values
\itemize{ 
\item \code{NA}: not born yet
\item \code{0}: susceptible 
\item \code{3}: exposed (infected but not infectious)
\item \code{1}: infectious
\item \code{9}: dead
}
}
}
